Add empty implementations of the core-to-plugin API
Adds empty implementations of the API that the LiveObjects plugin must
implement, as added in ably-cocoa commit dcca250. We'll implement these
methods, and refine their meaning, later.

Demonstrates how the calls to the plugin API find their way to the
DefaultLiveObjects object, and how we convert the ObjectMessage type
that we emit to ably-cocoa into the nice Swift struct that we'll use
internally.